requirements.txt
streamlit==1.44.1
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3